update advanced probplem
